Add files via upload
the initial value is taken as 3500
the periodicity is calculated and then files from 2022 are manually transferred to prediciting folder as the periodicity turns out to be 2. for a different set of stocks the periodicity may change and stocks to be moved to the  predicting folder. 
This model is preliminary and serves only as a recommendation. It advises against actual investment, as indicated by the model, and I am not held responsible for any losses incurred by following its suggestions